added image versioning via ci pipeline
The pipeline has to be run manually. It will build the docker image using the Dockerfile and push that image to the image repository. Then it will create a github release for this project and adjust its version.

The docker scripts have been adapted to use the image repository:
docker_run will now search for the fitting image locally and pull it from the image repository if not found.
docker_build can be used to build a fitting image locally, if the pulling via docker_run doesn't work for whatever reason.

A new version of the image is pushed to the image repo with every github release. This also means that there are no fitting image versions in the repo for checked out commits that don't match a release. More precisely, if the version in the VERSION file doesn't match a github release version, docker_run won't find the image. You can then either manually pull the image, using the last release version before your current state, or you build the image locally with docker_build.