kubectl aliases / cleanup script / colored bash prompt
The following aliases can now be used in the docker container:
ks => kubectl -n kube-system
kg => kubectl -n garden
ka => kubectl get --all-namespaces

Furthermore, the docker container now uses a colored prompt so that it is easier to differentiate output from commands.

The cleanup script will delete anything that is left over after the cluster is destroyed.
The script prompts for a confirmation, unless the "-y" option is given.