Split of README
To keep the documentation close to the corresponding code,
the readme has been splitted. One part, describing the project in general,
remains in the landscape-setup-template project.
The other part, describing the installation procedure, is moved into this project
as it describes the usage of the scripts located in this project.