separated image version from project version
Before, a new docker image was built with every release of this project.
docker_run.sh used the image tagged with the current project version.
However, the Dockerfile is assumed to change significantly less often than
other files of this project. To prevent unnecessary creation of identical
docker images with just different tags, there is now a IMAGE_VERSION file
that is independent from the project version and specifies the version of
the docker image that should be used.