added modular seed configuration
It is now possible to deploy several seeds on the cluster. There are several ways to do this:

1. On "deploy_gardener.sh", all seeds that are listed in seed_config.seeds in the landscape.yaml
are automatically deployed. On "undeploy_gardener.sh", all possible seeds are deleted to clean up
seed that were added manually (see below) and not leave any resources behind.

2. The seed-config deploy script now takes seeds as arguments. To create a seed, go to the
components folder and call './deploy.sh seed-config <seed>'. It is possible to provide multiple
seed arguments to create several seeds at once. If no arguments are given, the script uses the
seeds specified in the landscape.yaml, see above. Only the specified seeds are created, already
existing other seeds are not affected. This works the same way when the "-u"/"--uninstall" option
is given. The order of the arguments is not relevant.

Valid values for seeds are "aws", "az" (for Azure), "gcp", and "openstack". For every seed the
corresponding variants in the authentication and seed-config nodes in landscape.yaml have to exist
and be filled out.

Please note that while it is possible to create seeds for any cloud provider on any cloud provider,
actual shoot creation may not always work across the cloud providers. For example, creating an Azure
shoot from a seed on AWS works without problems, but creating a GCP shoot doesn't right now.