fixed: caching problem in read_landscape_config
The "read_landscape_config" function would only read the landscape config
if it isn't already cached or the name/path changed, using the cached config
otherwise. This sometimes lead to problems that changes in the config weren't
detected.
The function now also evaluates the config file's "last modified" flag
and reads it again if it differs from the time it was read last.