Fix certmanager deploy on single master setup
This fixes the certmanager's problem that a rolling update of the
kube-apiserver is not possible on a cluster with only one apiserver.
The solution is to reactivate the bootstrap apiserver that was used
to pull up the kube-apiserver in the first place. Then the patching
is done and afterwards the bootstrap apiserver is shut down again.
This happens automatically when the certmanager is deployed and
only one master is detected.