fix certmanager image version
The values file for the certmanager deployment now doesn't include
image name and version anymore.
The idea behind this is to modify the certmanager version by just
changing the argument given to helm and let certmanager itself decide
which image and version it wants to use (by using the default values
for these), as it should work best with this image.