add 'all' component
The 'all' component is a dummy component that replaces the old
deploy_gardener.sh and undeploy_gardener.sh scripts.
Both deploy and undeploy take one argument in the form of '<start>:<stop>'
and will (un)deploy components <start>, <stop>, and all components
in between. Partial arguments ('<start>:', ':<stop>', or no argument at
all) are also allowed and will use the defaults for the missing values.
The defaults are 'cert' and 'certmanager'.
The component order can be found in the environment variable
COMPONENT_ORDER. The order is reversed for the undeploy script.
Kubify as well as the etcd component are the only components not between
the default values, so an argument is needed if these should be
(un)deployed.
When using the script to deploy or undeploy kubify, it will skip the
etcd component, althouth it comes next in order after kubify. As this
Gardener deployment can integrate in the kubify etcd, the extra etcd
is not needed.