add 'all' component
The 'all' component is a dummy component that replaces the old
deploy_gardener.sh and undeploy_gardener.sh scripts.
Both deploy and undeploy take one argument in the form of '<start>:<stop>'
and will (un)deploy components <start>, <stop>, and all components
in between. Partial arguments ('<start>:', ':<stop>', or no argument at
all) are also allowed and will use the defaults for the missing values.
The defaults are the first and the last component of the current
component order variable.
There are multiple environment variables called COMPONENT_ORDER_<...>
and one of them will be used when the 'all' component is called. The
one to use is specified by the 'clusters.base_cluster' field in the
landscape.yaml file.
The order is reversed for the undeploy script.