add submodule management
Until now, there were submodules for gardener and dashboard as
well as a tag entry in the config file which version to use.
To prevent problems, both versions - the one specified in the
config file and the checked out version of the submodule - had
to match, so the information was redundant.
This commit removes this redundancy by removing the submodules.
A 'manage_submodules.sh' script is added that will create the
submodules, if they are not there and it will update them to
match the version specified in the config file.
The "charts" sections for gardener and dashboard now each contains
a "managed" field. If set to true, the automated submodule version
management is enabled for that submodule. The correct version will
be checked out during the deploy of that component.
Please note that the dashboard submodule will be updated also
during the "identity" component deploy, as that partly depends on
the dashboard submodule.