add submodule management
Until now, there were submodules for gardener and dashboard as
well as a tag entry in the config file which version to use.
To prevent problems, both versions - the one specified in the
config file and the checked out version of the submodule - had
to match, so the information was redundant.
This commit removes this redundancy by removing the submodules.
A 'manage_submodules.sh' script is added that will create the
submodules, if they are not there and it will update them to
match the version specified in the config file.
The script needs to be run manually, but it is needed only if
the submodules 'gardener' and 'dashboard' are not there yet
or if their tags in landscape_base.yaml changed.